Iframe changes to SSO (#1899)
* change(ui): iframe check for sso redirect

* change(ui): removed login.js

* change(ui): sso link

* change(ui): sso link

* change(ui): iframe check and sso redirect
